# backend/openapi.json
{"openapi": "3.1.0", "info": {"title": "FastAPI", "version": "0.1.0"}, "paths": {"/api/auth/signup": {"post": {"summary": "Signup", "operationId": "signup_api_auth_signup_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}, "409": {"description": "Conflict", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ExceptionModel"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/auth/login_challenge": {"post": {"summary": "Challenge", "operationId": "challenge_api_auth_login_challenge_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChallengeRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Challenge"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/auth/login_answer": {"post": {"summary": "Answer", "operationId": "answer_api_auth_login_answer_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Challenge"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}, "403": {"description": "Forbidden", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ExceptionModel"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/groups/create": {"post": {"summary": "Handle Create Group", "operationId": "handle_create_group_api_groups_create_post", "parameters": [{"name": "user_id", "in": "query", "required": true, "schema": {"type": "integer", "title": "User Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateGroupMessage"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Group"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/groups/add": {"post": {"summary": "Handle Add Group User", "operationId": "handle_add_group_user_api_groups_add_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/GroupAddUserRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/GroupMember"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/session/get_uid": {"post": {"summary": "Get Uid", "operationId": "get_uid_api_session_get_uid_post", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/GetUidResponse"}}}}}}}, "/api/users/": {"get": {"summary": "Handle Get User", "operationId": "handle_get_user_api_users__get", "parameters": [{"name": "username", "in": "query", "required": true, "schema": {"type": "string", "title": "Username"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"Challenge": {"properties": {"id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Id"}, "username": {"type": "string", "minLength": 8, "title": "Username"}, "challenge": {"type": "string", "title": "Challenge"}}, "type": "object", "required": ["username", "challenge"], "title": "Challenge"}, "ChallengeRequest": {"properties": {"username": {"type": "string", "title": "Username"}}, "type": "object", "required": ["username"], "title": "ChallengeRequest"}, "CreateGroupMessage": {"properties": {"type": {"type": "string", "const": "createGroup", "title": "Type"}, "name": {"type": "string", "title": "Name"}, "symmetric_key": {"type": "string", "title": "Symmetric Key"}}, "type": "object", "required": ["type", "name", "symmetric_key"], "title": "CreateGroupMessage"}, "ExceptionModel": {"properties": {"detail": {"type": "string", "title": "Detail"}}, "type": "object", "required": ["detail"], "title": "ExceptionModel"}, "GetUidResponse": {"properties": {"uid": {"type": "integer", "title": "Uid"}}, "type": "object", "required": ["uid"], "title": "GetUidResponse"}, "Group": {"properties": {"id": {"type": "integer", "title": "Id"}, "name": {"type": "string", "minLength": 1, "title": "Name"}, "owner_id": {"type": "integer", "title": "Owner Id"}}, "type": "object", "required": ["name", "owner_id"], "title": "Group"}, "GroupAddUserRequest": {"properties": {"user_id": {"type": "integer", "title": "User Id"}, "symmetric_key": {"type": "string", "title": "Symmetric Key"}, "group_id": {"type": "integer", "title": "Group Id"}}, "type": "object", "required": ["user_id", "symmetric_key", "group_id"], "title": "GroupAddUserRequest"}, "GroupMember": {"properties": {"user_id": {"type": "integer", "title": "User Id"}, "group_id": {"type": "integer", "title": "Group Id"}, "symmetric_key": {"type": "string", "title": "Symmetric Key"}}, "type": "object", "required": ["user_id", "group_id", "symmetric_key"], "title": "GroupMember"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "User": {"properties": {"id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Id"}, "username": {"type": "string", "minLength": 8, "title": "Username"}, "public_key": {"type": "string", "minLength": 1, "title": "Public Key"}}, "type": "object", "required": ["username", "public_key"], "title": "User"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}